feat: add Anaxa, Castorice, and Ruan Mei pages with navigation; update dashboard links and improve test imports
