+++ visual_dashboard/src/app/dashboard/castorice/page.tsx
@@ -1,4 +1,4 @@
-export default function AnaxaPage() {
+export default function CastoricePage() {
   return (
     <div className="flex flex-col items-center p-6 w-full h-full min-h-screen">
       <a
@@ -8,8 +8,8 @@ export default function AnaxaPage() {
         ← Back
       </a>
       <iframe
-        className="w-full h-[80vh] border-0"
-        src="https://lookerstudio.google.com/embed/reporting/d02f20cc-0710-4be7-87e2-ece693d00805/page/9R4IF"
+        className="w-[140vh] h-[80vh] border-0"
+        src="https://lookerstudio.google.com/embed/reporting/fefde0c5-1e9b-4e79-a3ba-ba036ee2b827/page/9R4IF"
         allowFullScreen
         sandbox="allow-storage-access-by-user-activation allow-scripts allow-same-origin allow-popups allow-popups-to-escape-sandbox"
         title="Anaxa Looker Studio Dashboard"
